refactor(#42): 앱 메시지 코드 파일 분리
