+++ src/components/auth/password/password-change-form.tsx
@@ -5,7 +5,7 @@ import AuthInputGroup from "@/components/auth/auth-input-group";
 import { Button } from "@/components/ui/button";
 import { FieldError } from "@/components/ui/field";
 import { Spinner } from "@/components/ui/spinner";
-import { APP_MESSAGE_CODE } from "@/constants/app-message";
+import { APP_MESSAGE_CODE } from "@/constants/app-message-code";
 import { createClient } from "@/lib/supabase/client";
 import { changePasswordSchema, ChangePasswordValues } from "@/lib/zod/auth";
 import { getAppMessageTitle } from "@/utils/app-message";
