+++ src/types/action.ts
@@ -1,6 +1,6 @@
 // 서버 액션의 공통 결과 타입을 정의하는 타입 모음
 
-import type { AppMessageCode } from "@/constants/app-message";
+import type { AppMessageCode } from "@/constants/app-message-code";
 
 export type AppActionResult<T = undefined> = {
   success: boolean;
